Add support for old EMXplus SPC files
Despite identical file extensions (SPC/PAR), the binary format of EPR300-E and (old) EMXplus spectrum files is different (mode and endianness). Because there is no way of determining the endianness programmatically, a new argument was introduced that lets the user specify the original device.
Thanks to Orlando Rodrigues Jr for bringing this to attention and for providing an example file.

NB: The "converter" shiny application was updated and now allows selecting the device, while also providing a new preview plot.